+++ pymongo/cursor.py
@@ -937,10 +937,11 @@ class Cursor(object):
         Can raise ConnectionFailure.
         """
         client = self.__collection.database.client
+
         try:
-            response = client._send_message_with_response(
-                operation, exhaust=self.__exhaust, address=self.__address,
-                unpack_res=self._unpack_response)
+            response = client._run_operation_with_response(
+                operation, self._unpack_response, exhaust=self.__exhaust,
+                address=self.__address)
         except OperationFailure:
             self.__killed = True